Cleanup events, start to prep for multi-db tests
